backend/requirements.txt
django-rest-framework==0.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
beautifulsoup4==4.12.2
requests==2.31.0